added error handling for errors durimg testing routes

+++ src/controllers/bill_controller.py
@@ -5,6 +5,7 @@ from init import db
 from datetime import datetime
 
 from flask_jwt_extended import jwt_required, get_jwt_identity
+from sqlalchemy.exc import ProgrammingError, DataError, StatementError
 
 
 bill_bp = Blueprint("bill", __name__, url_prefix= "/bill")
@@ -12,34 +13,37 @@ bill_bp = Blueprint("bill", __name__, url_prefix= "/bill")
 @bill_bp.route("/create", methods= ["POST"])
 @jwt_required()
 def create_bill():
-    body_data = BillSchema().load(request.get_json())
-    if all(body_data.get(field) for field in ["due_date", "amount", "bill_title"]):
-# Proceed if these 3 fields are present.
-        pass
-    else:
-        return {"Error": "Some fields are missing or empty = due_date, amount & bill_title must have data entered."}, 404
-    
-    bill = Bill(
-        due_date = body_data.get("due_date"),
-        amount = body_data.get("amount"),
-        bill_title = body_data.get("bill_title"),
-        description = body_data.get("description"),
-        paid = body_data.get("paid")
+    try:
+        body_data = BillSchema().load(request.get_json())
+        if all(body_data.get(field) for field in ["due_date", "amount", "bill_title"]):
+    # Proceed if these 3 fields are present.
+            pass
+        else:
+            return {"Error": "Some fields are missing or empty = due_date, amount & bill_title must have data entered."}, 404
         
-    )
-    stmt = db.select(Member).filter_by(member_id=get_jwt_identity())
-    member = db.session.scalar(stmt)
-    
-    if member:
-        bill.member_id = member.member_id
-        db.session.add(bill)
-        db.session.commit()
+        bill = Bill(
+            due_date = body_data.get("due_date"),
+            amount = body_data.get("amount"),
+            bill_title = body_data.get("bill_title"),
+            description = body_data.get("description"),
+            paid = body_data.get("paid")
+            
+        )
+        stmt = db.select(Member).filter_by(member_id=get_jwt_identity())
+        member = db.session.scalar(stmt)
+        
+        if member:
+            bill.member_id = member.member_id
+            db.session.add(bill)
+            db.session.commit()
 
-        return {
-        "Created Bill": bill_schema.dump(bill)
-        }, 201
-    else:
-        {"Error": "You do not have the authority to create Bills"}, 401
+            return {
+            "Created Bill": bill_schema.dump(bill)
+            }, 201
+        else:
+            return {"Error": "You do not have the authority to create Bills"}, 401
+    except (ProgrammingError, DataError, StatementError):
+        return {"Error": "Incorrect field format. Please enter correct format."}, 400
 
 @bill_bp.route("/", methods= ["GET"])
 @jwt_required()
@@ -74,28 +78,35 @@ def display_bills():
 @bill_bp.route("/update/<int:id>", methods= ["PUT", "PATCH"])
 @jwt_required()
 def update_bill(id):
-    body_data = BillSchema().load(request.get_json(), partial=True)
-    
-    stmt = db.select(Member).filter_by(member_id=get_jwt_identity())
-    member = db.session.scalar(stmt)
-    stmt2 = db.select(Bill).filter_by(bill_id = id)
-    bill = db.session.scalar(stmt2)
-
-    if bill and member:
-# Update the fields if they exist in the request data, or keep the current values
-        bill.due_date = body_data.get("due_date") or bill.due_date  
-        bill.amount = body_data.get("amount") or bill.amount       
-        bill.bill_title = body_data.get("bill_title") or bill.bill_title
-        bill.description = body_data.get("description") or bill.description
-        bill.paid = body_data.get("paid") or bill.paid
-        bill.member_id = member.member_id
+    try:
+        body_data = BillSchema().load(request.get_json(), partial=True)
         
-        db.session.commit()
-        return {
-            "Bill updated": "The bill fields have been updated."
-        }, 200
-    else:
-        return {"Error": "This bill does not exist or you dont have authority."}, 404
+        stmt = db.select(Member).filter_by(member_id=get_jwt_identity())
+        member = db.session.scalar(stmt)
+        group_id = member.fam_group_id
+        stmt2 = db.select(Bill).filter_by(bill_id = id)
+        bill = db.session.scalar(stmt2)
+        
+        if bill.member.fam_group_id != member.fam_group_id:
+            return {"Error": "Update of Bills not in your same family group is not allowed. "}
+
+        if bill and member:
+    # Update the fields if they exist in the request data, or keep the current values
+            bill.due_date = body_data.get("due_date") or bill.due_date  
+            bill.amount = body_data.get("amount") or bill.amount       
+            bill.bill_title = body_data.get("bill_title") or bill.bill_title
+            bill.description = body_data.get("description") or bill.description
+            bill.paid = body_data.get("paid") or bill.paid
+            bill.member_id = member.member_id
+            
+            db.session.commit()
+            return {
+                "Bill updated": "The bill fields have been updated."
+            }, 200
+        else:
+            return {"Error": "This bill does not exist or you dont have authority."}, 404
+    except (ProgrammingError, DataError, StatementError):
+        return {"Error": "Incorrect field format. Please enter correct format."}, 400
 
 
 @bill_bp.route("/delete/<int:id>", methods= ["DELETE"])
@@ -105,8 +116,12 @@ def delete_bill(id):
     bill = db.session.scalar(stmt)
     stmt2 = db.select(Member).filter_by(member_id = get_jwt_identity())
     member = db.session.scalar(stmt2)
+    
     if not bill:
         return {"Error": f"Bill with id: {id}, does not exist."}, 404
+    if bill.member.fam_group_id != member.fam_group_id:
+            return {"Error": "Deletion of Bills not in your same family group is not allowed. "}
+    
     if member.is_admin and member:
         db.session.delete(bill)
         db.session.commit()
